+++ account/urls.py
@@ -1,0 +1,17 @@
+from django.urls import path
+from . import views
+
+# allauth 
+from django.urls import path, include, re_path
+from django.views.generic import TemplateView
+from django.contrib.auth.views import LogoutView
+
+urlpatterns=[
+    # Login and registration page 
+    path("register_login", views.register_login, name="register_login"),
+    # all auth views
+    path('accounts/', include('allauth.urls')),
+    # My Account page
+    path("myaccount", views.myaccount, name="myaccount"),
+    path('logout', LogoutView.as_view()),
+]
