+++ doobarashop/views.py
@@ -24,21 +24,8 @@ def filtering(request, locat):
     slop = [row.serialize() for row in lop]
     return render(request, "doobarashop/shop.html", {"lop":slop})
 
-
-def blog(request):
-    return render(request, "doobarashop/blog.html")
-
-def video(request):
-    return render(request, "doobarashop/video.html")
-
 def contactus(request):
     return render(request, "doobarashop/contactus.html")
-
-def myaccount(request):
-    # if request.user.is_authenticated:
-    return render(request, "doobarashop/account.html")
-    # return render(request, "account/login.html")
-
 
 
 def single_product(request, locat):
@@ -46,11 +33,9 @@ def single_product(request, locat):
     product = product.serialize()
     return render(request, "doobarashop/single_product.html", {"product": product})
 
-def single_blog_post(request):
-    return render(request, "doobarashop/single_blog_post.html")
 
-def register_login(request):
-    return render(request, "doobarashop/register_login.html")
+
+
